Simulation: Use parameters method
Before this commit, the parameters method was implemented by all the
simulation mixins, but went unused. This resulted in confusing behaviour.
Furthermore, the various mixins would set parameters in their own
overwritten implementations of initialize().

With this commit, simulation child classes and mixins can override the
parameters() method to set parameters instead of needing to implement
and maintian their own set_var() logic. This also makes it much more
transparent to the user what the values of the parameters are and which
class takes precedence when setting parameters.