Fix parameter inlining for delayed expressions
If we inline a parameter, we have to make sure to also inline it in the
delay expression and duration. The latter was already taken care of
somewhere else in the code, but has been moved near the newly added
processing of the delay expressions.

Note that evaluating the delayed_feedback_function() could still
give non-NaN results even if inputs were missing, which made this bug
somewhat hard to encounter and track down.
The fact that this happens is due to the expand() call; if we would call
the original MX function, it would complain about free variables. Why
CasADi does not always return NaN for outputs of which not all inputs
were specified is unknown though.