PIMixin: Support not exporting initial state
The results at the initial time can contain bogus values for (algebraic)
states that do not have their initial state specified. Note that not
specifying an initial state for algebraic states may be desired to avoid
infeasibilities.
    When running a closed loop hindcast with FEWS, these bogus values
mean that the user will have to do a transformation for every imported
time series to cut off the first value (lest it overwites a correct
existing value).

This commit makes that process much easier, by allowing users to set a
flag to make PIMixin skip writing the results at the initial time.

Closes #1049