New goal_programming_mixin
If the option 'keep_eps_variable' is set to False, the old
implementation of the goal programming mixin is used. This is the
default option.

If 'keep_eps_variable' is set to True, the epsilon variables are
kept throughout the optimization runs.
At each new optimization run a constraint enforcing Pareto
optimality is added.

If the goal programming option 'linear_obj_eps' is set to True, then
each optimization problems will have a linear objective.
This option can be
used only when all the minimization goals have order one.
To deal with goal with targets and order different from one, alias
epsilon variables are introduced.