Fix goal relaxation with minimization goals
The default value for fix_minimized_values is True, and therefore a goal
relaxation on minimization goals would result in two constraints, one of
which would be possibly non-convex.

The idea behind fix_minimized_values is to impose a possibly non-convex
_equality_ constraint for interior point methods, to avoid numerical
issues with small slack variables. We therefore now check if the goal
has any relaxation, in which case we impose a single inequality
constraint instead of a single equality constraint.