Add options to force linear objective for goals
Some solvers are not able to handle problems with both quadratic
constraints _and_ a quadratic objective. They do however support a linear
objective together with quadratic constraints. In this commit an option
``force_linear_objective`` is added that makes it easier for the user to
use solvers like this when e.g. order=2 goals are desired.

The objective is linearized by adding additional variables and constraints
to the problem. For example, if the objective function would be equal to
eps_1^2 + eps_2^2, then it will become eps*_1 + eps*_2 with the additional
constraints eps*_1 >= eps_1^2 and eps*_2 >= eps_2^2.