Add 'objective_constraint_relaxation' option
Add the goal programming option obj_constr_relaxation. This
allows to relax the constraint arising from the objective function at
each priority by a specified value.

It can either a single value or a list of values. If it is a list, there
must be the same number of values as the number of active objective
constraints (i.e., number of active priorities minus one).

Not yet implemented for keep_soft_constraints False.