Add history to CSVMixin
Adds a csv_forecast_date variable to CSVMixin. The part of the
input timeseries until this forecast date is considered historic. The
initial time (t0) is set to this forecast date. The output timeseries
will only contain the times from the forecast date onwards. The historic
values will be used for delay variables. By default the forecast date is
the start time of the input timeseries.