Seed force_linear_objective variables
Seed the variables used by the 'force_linear_objective' option at the
current priority.

If there is a corresponing epsilon varible, the seed is equal to one.
Else, the solver_ouput is used to compute a value for variable. This
ensures that a feasible starting point is provided to the solver.