Allow goal relaxation with keep_soft_constraints
If a goal has target bounds, after the optimization run of the priority
for the goal, the constraints are overwritten with a relaxation
corresponding to goal.relaxation / goal.nominal.

If a goal is a minimization one with a positive goal relaxation, a
separate objective constraint is defined for this goal and it is
relaxed by goal.relaxation / goal.nominal.

All the ojective functions of the remaining goals are aggregate to form
a single constraint.

All the constraints related to the objective functions are relaxed with
the constraint_relaxation.