Add general DataStore and IOMixin classes
Large refactoring of all IO mixins (PIMixin and CSVMixin).
Adds a general DataStore class extended by OptimizationProblem
and SimulationProblem, which is used to store all data read by
the IO mixins.
Adds an optimization IOMixin class which contains methods that
were previously duplicated in the optimization PIMixin and
CSVMixin.
Adds a simulation IOMixin class which does the same for the
simulation PIMixin and IOMixin.