SimulationProblem: Fix scaling of equations with nominals
SimulationProblem was only ever using the nominals defined in the Modelica
file, even if e.g. the user was overriding get_variable_nominal().

Also note that variable nominals can be symbolic, but that we would like
to handle numeric values/floats as much as possible. We therefore
evaluate the nominals to their numeric values as soon as we know all
parameter values.