SimulationProblem: Fix when nominal is symbolic MX
Due to backwards compatibility with allowing set_var() for parameters,
it is possible for get_variable_nominal() to return a symbolic MX
expression instead of a numeric value. This was not handled properly
inside SimulationProblem.

Note that we want to disallow setting parameters like this in the
future, instead forcing users to override/extend the parameters()
method. That way, we will be able to guarantee that
get_variable_nominal() always returns a numeric value.

Closes #1085