Support absolute goal order for minimization goals
One would typically use order=2 to minimize the absolute value of the goal
function, but that makes nominals a bit harder and easily leads to scaling
issues. A more involved option that was used is to introduce an auxiliary
variable and additional constraints to minimize the absolute value of the
goal function.

It is the latter logic that this commit adds to GoalProgrammingMixin, such
that the user now only has to set order="abs" on a Goal instance. The
auxiliary variable and accompanying constraints are automatically added
behind the scenes.

Closes #1065