GPMixin: Add optional linearization of goal order
For some problems it is preferred to not introduce any non-linear terms
in the objective. For example, when using a MILP solver, one wants the
entire problem to remain linear. This sometimes clashed with the desire
for an order 2 goal, which steers the solution in the direction of many
small exceedences instead of one large exceedence. An order 1 goal
does not distuinguish between the two.

This commit adds functionality to linearly approximate the goal order.
That way, one can keep the problem fully linear, why getting some of the
benefits of higher order goals back.

Aside from solvers which only handle LP and not QP, using a linearly
approximated goal order is also useful when `keep_soft_constraints` is
set. A quadratic problem in priority 1 would mean a quadratically
_constrained_ problem in priority 2 with that option, something that can
be much harder to solve.

Note that higher order minimization goals do not work with
`linearize_goal_order`. This is because a function range is lacking over
which to linearize. Instead, users are requested to give these
minimization goals e.g. a target_min = target_max = 0.0 (and a proper
function range), such that they can be linearized. Minimization goals
with order 1 are supported, as they do not have to be linearized.