ModelicaMixin: Fix reading of symbolic initial states
The initial_state() method has to return a numeric value. For symbolic
initial state that means that we need to substitute the parameters, just
like e.g. bounds() and seed() already did.

Note that, unlike e.g. the seed() method in ModelicaMixin, we raise an
exception if the symbolic expression cannot be resolved. This is because
a wrong/missing initial state is worse than a wrong seed value.