pi.py: Fix duplicating read values for all ensembles
When an ensemble member is not specified for a certain variable in the
input, we want to add the same variable-values pair to all ensembles.
The loop that was supposed to do this was a no-op, as it used
`ensemble_member` (which is zero in this case), instead of the loop
index.

Note that we remove the underscore prefix that indicated that the loop
index was unused.