PIMixin: Fix overriding initial state values
If a user was using ModelicaMixin and PIMixin, the initial state values
from ModelicaMixin were skipped. We have to be sure to call the respective
super() method. Note that CSVMixin's initial_state() also calls super().

Also note that ModelicaMixin's initial_state() does not call super(), to
avoid getting stuck in an endless loop due to OptimizationProblem's
implementation of initial_state() and history(). This is a separate bug,
which cannot easily be addressed without significantly refactoring the way
history and initial states are stored.

Closes #1110