GPMixin: Fix hard constraint for minimization goals
Make sure to take the goal order into account when constructing a hard
constraint for minimization goals. This reformulation allows for a
larger feasible solution space and a proper application of
a goal's relaxation.

Note that this commit changes nothing for goals with target bounds. It
also only applies in the case that keep_soft_constraints is False.