GPMixin: Fix array targets for vector goals
When a np.ndarray was specified for a target min or max, it would be
interpreted the wrong way. A test has been modified to check this
behavior.

The fix is related to the internally used method _gp_min_max_arrays(),
which was very inconsistent in the shape of the data it would return. It
has been fixed to always:

- Output a 2-D array of size (goal.size, target_shape or 1) if the goal
  size is larger than one, i.e. a vector goal.
- Output a 1-D array of size (target_shape or 1, )

Fixing this consistency exposed some (potential or known) bugs elsewhere
in the code, which have also been addressed.