GPMixin: Fix division by zero for vector goal with scale_by_problem_size
If a certain element of a vector goal is inactive at every time step, it
would result in n_active being zero. When scale_by_problem_size is set,
we divide by this value, which would then result in an error. To fix
this, we clip the minimum value of n_active to 1.