run_optimization_problem: Allow passing arguments to problem class
BACKPORT NOTE: This is a backported feature, not a bug fix. It turns out
that passing information to the optimization/simulation problem is a
common desire/requirement, e.g. passing along commandline arguments. Up
until this commit, users were forced to use global variables to pass
such information, and we definitely do not want to encourage the usage
of global variables. We therefore backport the feature added in this
commit such that users can properly pass information as arguments.

The functions run_optimization_problem() and run_simulation_problem()
are commonly used to instantiate and run models. Sometimes it can be
useful to override some default instance/class variables on the problem
class. This commit makes that much easier by allowing one to pass
additional keyword arguments to run_optimization_problem() and
run_simulation_problem(), which are then passed through to the problem
class's __init__().

Note that additional positional arguments are not allowed for future
compatibility. We may want to pass additional arguments ourselves to
a problem class in the future.