CollInt: Fix interpolation in state_at for integrated variables
The length of times should match the length of values. However, the
zeroth state was missing from the list of values. Because
OptimizationProblem's __interpolate() does not check on this, it still
succeeded. Note that we avoid making a check in this commit, because we
should not be calling self.interpolate() for MX symbols to begin with.
That will be solved in a separate commit.