IOMixin: Fix setting timeseries with subset of import times
We raised an Exception when calling set_timeseries where the times were
a subset of the import times. This is a regression in 623fc2cf (IO
refactor). We should allow such cases, and just stretch the values. Note
that this goes both for calling set_timeseries with a Timeseries, and
for calling it with a np.array.