CollInt: Simplify extract_states() and extract_controls()
Much of the extraction logic for collocated variables, path variables
and extra variables was the same. We now group the processing of these
variables to reduce duplicate code.

Note that relying on indices makes the implementation of
extract_controls() in ControlTreeMixin superfluous. Relying on the base
implementation works just as well.