Tests: Fix `test_differences_in_timesteps` test
`NetCDF4` depends on `cftime`. For some reason, older versions of cftime
had some rounding errors when handling arrays. For the mentioned test,
when comparing the NetCDF file to the PI-XML file, we inadvertently
relied on this behavior. In actuality, both files have the same
datetimes.

This commit makes sure that the three input files (CSV, PI-XML, and
NetCDF) have different datetimes. The CSV file has offsets of 0-5
minutes, the NetCDF file 0-5 hours, and the PI-XML now has 0-5 seconds.

See also the following issues/PRs on `cftime` which exposed this:
https://github.com/Unidata/cftime/issues/144
https://github.com/Unidata/cftime/pull/146