run_*_problem: Workaround `Class not found` for Spyder/Embeddable Python
Both Spyder and Embeddable Python do not set sys.path[0] to its correct
value of the folder containing the script that is executed. Instead of
asking people to amend their script to insert the correct folder in
sys.path, we can try a little harder ourselves.

For examples, Spyder has a habit of giving the complete/absolute path to
the executed script (stored in `sys.argv[0]`). With a relative
`base_folder` path, we can then easily figure out the absolute path.

For Embeddable Python, the path to the script in `sys.argv[0]` is
relative. The remaining thing we can try is to resolve the path to the
script relative to the current working directory. This assumes that
people do not change working directory in weird ways, which is why we
only try this as a last resort. Note that we also assume people do not
not manipulate sys.path in weird ways before calling e.g.
`run_optimization_problem`, but that was already the case before this
commit.