Style: Reformat code using black (line length = 100)
We also add black to the CI/CD pipeline to check future changes.

Note that black makes certain changes that flake8 disagrees with (most
notably E203). We disable that check, along with two other conventions
we like.

Also note that we keep the line length for flake8 at 120 for now. This
is because there are many comments and docstrings that are longer than
100 characters. The intention is to make a separate commit addressing
those.