Simulation: Substitute modelica parameters in functions
One can pass to Ipopt parameters or do the substitution before hand. We
now do the latter, to allow to solve for equations of the type x^k
where k is a parameter.

Note that this is not an issue in optimization problem, as in CollIntOpt
the parameter substitution is already done.

TODO:
- allow parameters to be able to be overwritten (not coming directly
  from modelica)
- (remove the style commit)