CollIntOptimizationProblem:  Fix and modernise DAE integration functionality
Before, there were no tests combining integrated states with goal programming.  Several pieces of functionality, such as path variables, goals, and so forth, were not being tested in combination with integrated states.  As a result, changes were made to the goal programming implementation (as well as to the CollInt core) that broke these combinations.  This PR introduces, besides the fixes, new tests for the previously broken combinations.

Integrating "some" states works when the number of model equations that depends on the integrated states is equal to the number of integrated states.  Such an equality would enable one to decompose the model into integrated and non-integrated parts.  In general, however, there will be cross-dependencies and for this reason the equality will not hold.  This PR simplifies the logic such that either all states are integrated, or none.

Integration is only applicable to model states.  Path variables are auxiliary variables at the level of the optimization problem, i.e., are introduced at a different level.  The model sits at the core of RTC-Tools; around the core, it offers functionalities such as goal programming which wrap around, but do not modify, the model.