simulation: update writing variables to xml file
When writing a variable to an xml file,
it will loop over all aliases of this variable.
The first alias that is in the rtcDataConfig file will be written
to the xml file.
The order of the aliases is random and therefore the actual
variable that will be written is random.
By ensuring that the original variable appears first in the list
of aliases the behaviour is a bit less random and unexpected,
especially when the variable itself is in the variable file.