Revert "simulation: change time-step solver to fast_newton"
This reverts commit d2c97f4789ced0f1f492d957ade94e6da0032ff4.

The fast_newton solver fails to find a solution for certain models.
To prevent current use cases from suddenly breaking,
we return to the previous solver.
In the future, an option will be added to SimulationProblem
with which the user can select his/her own solver for the implicit
time step updates.