CI/CD: Run examples job in parallel with coverage
The examples are only run for tags, we can therefore be rather sure that
most stages of the pipeline will succeed. To save a little bit of time,
we can run the examples job in parallel with the other tests and
coverage, especially because the examples take quite a while to run.

Note that we add coverage as an explicit dependency to the deploy job,
and as such sill require all jobs to succeed before deploying like.