CI/CD: Create deploy job
Adding a separate workflow for deploying is possible, but to _also_ make
it depend on the tests succeeding is difficult. We would have to use the
`workflow_run` event, but that one cannot filter on tags, only on
branches. As such, we might as well add it to the overall workflow.